backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg[binary]
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt<4.0
python-multipart
email-validator
pytest
pandas
numpy
scikit-learn
torch
transformers
sentence-transformers
faiss-cpu
nltk
matplotlib
joblib
httpx
openai
